♻️ extract shared callback logic to base class
Extract common agent callback functionality into BaseAgentCallbacks base class to eliminate code duplication between docker-swarm and bedrock-agents implementations.

Changes:
- Create BaseAgentCallbacks parent class with shared functionality
- Remove duplicate imports (boto3, json, os, etc.) from child classes
- Consolidate common methods (__init__, metadata property, reset_metadata)
- Add helper method _extract_tool_parameters to reduce repeated code
- Update AgentCallbacks to inherit from BaseAgentCallbacks
- Maintain same external interface and behavior
