+++ lib/agent-core/docker/src/callbacks.py
@@ -5,15 +5,9 @@
 # ------------------------------------------------------------------------ #
 from __future__ import annotations
 
-import json
-import os
 from typing import TYPE_CHECKING
 
-import boto3
-from botocore.exceptions import ClientError
-from shared.base_callbacks import FormatCitations
-from shared.base_constants import RETRIEVE_FROM_KB_PREFIX
-from shared.kb_types import Citation, RetrievedReference
+from shared.base_callbacks import BaseAgentCallbacks
 from strands.hooks.events import (
     AfterToolCallEvent,
     BeforeToolCallEvent,
@@ -22,38 +16,28 @@ from strands.hooks.events import (
 if TYPE_CHECKING:
     from logging import Logger
 
-# --------------------------- AWS CLIENTS ---------------------------- #
-SNS_CLIENT = boto3.client("sns", region_name=os.environ.get("AWS_REGION"))
-AGENT_TOOLS_TOPIC_ARN = os.environ.get("agentToolsTopicArn", "")
-# -------------------------------------------------------------------- #
 
-
-class AgentCallbacks:
+class AgentCallbacks(BaseAgentCallbacks):
     """Class to handle callbacks for agent operations.
 
-    This class provides callback handlers for various agent operations including tool invocations
-    and knowledge base retrievals. It manages metadata collection and logging during agent execution.
+    This class extends BaseAgentCallbacks with trajectory enrichment features
+    for evaluation purposes. It stores tool arguments and results keyed by tool_call_id
+    to enable post-processing of agent trajectories.
 
     Attributes:
         _metadata (dict): Dictionary storing metadata generated by tools, e.g. Knowledge base references.
         _logger (Logger): Logger instance for recording agent operations and tool usage.
+        _tool_executions (dict): Dictionary mapping tool_call_id to tool execution data.
     """
 
     def __init__(self, logger: Logger, session_id: str, user_id: str) -> None:
-        self._metadata = dict()
-        self._logger = logger
-        self._nb_tool_invocations = 0
-        self._session_id = session_id
-        self._user_id = user_id
-
-    @property
-    def metadata(self) -> dict:
-        return self._metadata
+        super().__init__(logger, session_id, user_id)
+        self._tool_executions: dict[str, dict] = {}
 
     def reset_metadata(self) -> None:
-        self._metadata = dict()
-        self._nb_tool_invocations = 0
-        self._tool_executions: dict[str, dict] = {}
+        """Reset metadata, tool invocation counter, and tool executions."""
+        super().reset_metadata()
+        self._tool_executions = {}
 
     @property
     def tool_executions(self) -> dict[str, dict]:
@@ -91,8 +75,6 @@ class AgentCallbacks:
         tool_input = event.tool_use.get("input", {})
 
         if tool_call_id:
-            if not hasattr(self, "_tool_executions"):
-                self._tool_executions = {}
             self._tool_executions[tool_call_id] = {
                 "name": tool_name,
                 "arguments": tool_input,
@@ -102,68 +84,13 @@ class AgentCallbacks:
                 extra={"toolName": tool_name, "toolInput": tool_input},
             )
 
-        parameters = []
-        input_values = event.tool_use.get("input", {})
+        # Extract parameters using base class helper
+        parameters = self._extract_tool_parameters(event, specs)
 
-        if specs:
-            input_schema = specs.get("inputSchema", {}).get("json", {})
-            properties = input_schema.get("properties", {})
-
-            for param_name, param_value in input_values.items():
-                param_schema = properties.get(param_name, {})
-                parameters.append(
-                    {
-                        "name": param_name,
-                        "type": param_schema.get("type", "unknown"),
-                        "description": param_schema.get("description", ""),
-                        "value": param_value,
-                    }
-                )
-        else:
-            # Fallback if no specs available
-            for param_name, param_value in input_values.items():
-                parameters.append(
-                    {
-                        "name": param_name,
-                        "type": "unknown",
-                        "description": "",
-                        "value": param_value,
-                    }
-                )
-
-        message = {
-            "context": {
-                "userId": self._user_id,
-                "sessionId": self._session_id,
-                "invocationNumber": self._nb_tool_invocations,
-            },
-            "data": {
-                "toolName": event.tool_use.get("name", "unknown"),
-                "toolDescription": specs.get("description", "") if specs else "",
-                "parameters": parameters,
-            },
-        }
-
-        # Publish tool invocation to SNS topic
-        if AGENT_TOOLS_TOPIC_ARN:
-            try:
-                SNS_CLIENT.publish(
-                    TopicArn=AGENT_TOOLS_TOPIC_ARN,
-                    Message=json.dumps(message),
-                    Subject="AgentToolInvocation",
-                )
-                self._logger.info(
-                    "Published tool invocation to SNS",
-                    extra={
-                        "topicArn": AGENT_TOOLS_TOPIC_ARN,
-                        "toolName": message.get("toolName"),
-                    },
-                )
-            except (ClientError, ValueError) as err:
-                self._logger.warning(
-                    "Failed to publish tool invocation to SNS",
-                    extra={"error": str(err)},
-                )
+        # Publish tool invocation using base class helper
+        tool_name = event.tool_use.get("name", "unknown")
+        tool_description = specs.get("description", "") if specs else ""
+        self._publish_tool_invocation(tool_name, tool_description, parameters)
 
     def log_tool_results(self, event: AfterToolCallEvent) -> None:
         """Logs tool results and stores them for trajectory enrichment.
@@ -207,71 +134,9 @@ class AgentCallbacks:
                 result_content = str(event.result)
 
             # Update the stored tool execution with result
-            if (
-                tool_call_id
-                and hasattr(self, "_tool_executions")
-                and tool_call_id in self._tool_executions
-            ):
+            if tool_call_id and tool_call_id in self._tool_executions:
                 self._tool_executions[tool_call_id]["result"] = result_content
                 self._logger.debug(
                     f"Stored tool result for trajectory enrichment: {tool_call_id}",
                     extra={"toolName": tool_name, "resultLength": len(result_content)},
                 )
-
-    def retrieve_from_kb_callback(self, event: AfterToolCallEvent) -> None:
-        """Callback handler for knowledge base retrieval operations.
-
-        Processes retrieval results from knowledge base queries, formats citations,
-        and updates metadata with references.
-
-        Args:
-            event (AfterToolInvocationEvent): Event containing tool invocation results
-                and metadata.
-
-        Returns:
-            None
-        """
-        if event.selected_tool and event.selected_tool.tool_name.startswith(
-            RETRIEVE_FROM_KB_PREFIX
-        ):
-            sources = event.result.get("content", [{}])[0].get("json", {}).get("retrievalResults", [])  # type: ignore
-            if not sources:
-                # No sources found --> return
-                return
-
-            citations = Citation(
-                retrievedReferences=[
-                    RetrievedReference.model_validate(s) for s in sources
-                ]
-            )
-            formatter = FormatCitations([citations])
-
-            # Get existing unique references (RetrievedReference objects)
-            existing_unique_refs = self._metadata.get("_unique_references", [])
-
-            # Combine with new ones using set to deduplicate (leveraging __hash__ and __eq__)
-            all_unique_refs = list(
-                set(existing_unique_refs + list(formatter._unique_references))
-            )
-
-            # Store the unique references for future deduplication
-            self._metadata["_unique_references"] = all_unique_refs
-
-            # Format references with sequential IDs
-            references = []
-            for ref_id, ref in enumerate(all_unique_refs):
-                references.append(
-                    {
-                        "referenceId": str(ref_id + 1),
-                        "uri": ref.location.get_id(),
-                        "pageNumber": str(ref.get_page_number()),
-                        "content": ref.content.get_content(),
-                        "documentTitle": (
-                            ".".join(ref.metadata["documentName"].split(".")[:-1])  # type: ignore
-                            if "documentName" in ref.metadata
-                            else ref.location.get_id()
-                        ),
-                    }
-                )
-
-            self._metadata["references"] = references
